Apply "64bit fix" patch to fix the warnings on 64-bit machines. (sadly, does not fix the statuscolors pointer abuse warnings)
git-svn-id: svn://sporkhack.com/nh@510 0b9a868d-f130-0410-bc2a-adc7ae8015a6

+++ src/apply.c
@@ -960,7 +960,7 @@ struct obj* obj;
 	check_unpaid(obj);
 	pline("%s %s burns with a dim flame and vile stench.", Shk_Your(buf, obj), xname(obj));
 	begin_burn(obj, FALSE);
-	BStealth = (int)obj;
+	BStealth = (size_t)obj;
 
 }
